Moved RPC calls to service folder
