+++ web-client/src/content/Login.tsx
@@ -7,7 +7,6 @@ import TextField from '@material-ui/core/TextField'
 import React, { useEffect, useReducer } from 'react'
 import { useNavigate } from 'react-router-dom'
 
-import { enroll, getClasses, getEnrollment } from '../client'
 import LoginButton from '../components/LoginButton'
 import ScreenTemplate from '../components/ScreenTemplate'
 
@@ -88,9 +87,6 @@ function Login() {
     const [state, dispatch] = useReducer(reducer, initialState)
   
     useEffect(() => {
-        getClasses().then()
-        getEnrollment().then()
-        enroll().then()
         if (state.username.trim() && state.password.trim()) {
             dispatch({
                 type: 'setIsButtonDisabled',
